LayoutTests/inspector/css/variablesForType.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script>

function test()
{
    let suite = InspectorTest.createAsyncSuite("CSSStyleDeclaration.variableStylesByType");

    function addTestCase({name, description, selector, expected})
    {
        suite.addTestCase({
            name,
            description,
            async test() {
                let documentNode = await WI.domManager.requestDocument();
                let nodeId = await documentNode.querySelector(selector);
                let domNode = await WI.domManager.nodeForId(nodeId);
                InspectorTest.assert(domNode, `Should find DOM Node for selector '${selector}'.`);

                let domNodeStyles = WI.cssManager.stylesForNode(domNode);
                await domNodeStyles.refreshIfNeeded();

                domNodeStyles.computedStyle.useVariablesGrouping = true;

                for (let {groupType, variables} of expected) {
                    let propertyNames = domNodeStyles.computedStyle.variablesForType(groupType).map(property => property.name);
                    InspectorTest.expectTrue(!!propertyNames.length, `Should have "${groupType}" variables group.`);
                    InspectorTest.expectShallowEqual(propertyNames.sort(), variables.sort(), "Group should include all expected variables.");
                }
            },
        });
    }

    addTestCase({
        name: "CSSStyleDeclaration.variableStylesByType.Colors",
        description: "Check that variables with color values are correctly identified.",
        selector: "#colors",
        expected: [
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Colors,
                variables: ["--colorName", "--colorHEX", "--colorRGB", "--colorRGBA", "--colorHSL", "--colorHSLA"],
            },
        ],
    });

    addTestCase({
        name: "CSSStyleDeclaration.variableStylesByType.Dimensions",
        description: "Check that variables with dimension values are correctly identified.",
        selector: "#dimensions",
        expected: [
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Dimensions,
                variables: ["--dimensionPX", "--dimensionPercent", "--dimensionFloat", "--dimensionNegative", "--dimensionNegativeFloat"],
            },
        ],
    });

    addTestCase({
        name: "CSSStyleDeclaration.variableStylesByType.Numbers",
        description: "Check that variables with number values are correctly identified.",
        selector: "#numbers",
        expected: [
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Numbers,
                variables: ["--number", "--numberFloat", "--numberNegative", "--numberNegativeFloat"],
            },
        ],
    });

    addTestCase({
        name: "CSSStyleDeclaration.variableStylesByType.Other",
        description: "Check that variables with values with unmatched types are placed in the fallback group.",
        selector: "#other",
        expected: [
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Other,
                variables: ["--shadow", "--font-family", "--font-shorthand", "--keyword"],
            },
        ],
    });

    addTestCase({
        name: "CSSStyleDeclaration.variableStylesByType.Mixed",
        description: "Check that all type groups have variables allocated.",
        selector: "#mixed",
        expected: [
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Colors,
                variables: ["--color"],
            },
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Dimensions,
                variables: ["--dimension"],
            },
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Numbers,
                variables: ["--number"],
            },
            {
                groupType: WI.CSSStyleDeclaration.VariablesGroupType.Other,
                variables: ["--font-family"],
            },
        ],
    });

    suite.runTestCasesAndFinish();
}
</script>
<style>

div#colors {
    --colorName: black;
    --colorHEX: #000;
    --colorRGB: rgb(0, 0, 0);
    --colorRGBA: rgba(0, 0, 0, 1);
    --colorHSL: hsl(0, 0%, 0%);
    --colorHSLA: hsla(0, 0%, 0%, 1);

    /* should not match any below */
    --shadow: 1px 1px black;
    --dimensionPX: 100px;
    --number: 100;
}

div#dimensions {
    --dimensionPX: 100px;
    --dimensionPercent: 100%;
    --dimensionFloat: 1.0%;
    --dimensionNegative: -100px;
    --dimensionNegativeFloat: -1.0%;

    /* should not match any below */
    --shadow: 1px 1px black;
    --color: black;
    --number: 100;
    --numberFloat: 1.0;
    --numberNegativeFloat: -1.0;
}

div#numbers {
    --number: 100;
    --numberFloat: 1.0;
    --numberNegative: -100;
    --numberNegativeFloat: -1.0;

    /* should not match any below */
    --shadow: 1px 1px black;
    --dimensionPX: 100px;
    --dimensionPercent: 100%;
    --dimensionNegative: -100px;
    --dimensionNegativeFloat: -1.0%;
}

div#other {
    --shadow: 1px 1px black;
    --font-family: sans-serif;
    --font-shorthand: 16px / 1.6 sans-serif;
    --keyword: inherit;
}

div#mixed {
    --color: black;
    --dimension: 100px;
    --number: 100;
    --font-family: sans-serif;
}

</style>
</head>
<body onload="runTest();">
<p>Test for CSSStyleDeclaration.variableStylesByType</p>
<div>
    <div id="colors"></div>
    <div id="dimensions"></div>
    <div id="numbers"></div>
    <div id="other"></div>
    <div id="mixed"></div>
</div>
</body>
</html>
